+++ src/display_error_damaged_media.c
@@ -24,7 +24,7 @@ void display_error_damaged_media(void)
         in_error = 1;
         if ( we_are_at_the_main_menu )
         {
-            if ( widget_globals.deferred_error[0].error_code == 0xFFFF )
+            if ( (uint16_t)widget_globals.deferred_error[0].error_code == 0xFFFF )   /* DEVIATION: int16_t field — uncast, the promoted -1 never matched, so an already-queued error was silently overwritten; binary zero-extends, lhz _widget_globals.deferred_error(r31) @0x837310E8 + cmplwi cr6,r11,0xFFFF @0x837310EC */
             {
                 widget_globals.deferred_error[0].error_code = _error_media_damaged;
                 widget_globals.deferred_error[0].local_player_index = 0;
@@ -34,7 +34,7 @@ void display_error_damaged_media(void)
         }
         else
         {
-            if ( widget_globals.main_menu_deferred_error_code == 0xFFFF )
+            if ( (uint16_t)widget_globals.main_menu_deferred_error_code == 0xFFFF )   /* DEVIATION: same shape — lhz _widget_globals.main_menu_deferred_error_code(r31) @0x837310CC + cmplwi cr6,r11,0xFFFF @0x837310D0 */
                 widget_globals.main_menu_deferred_error_code = _error_media_damaged;
             main_goto_main_menu();
         }
